+++ sierra_typesafe/src/main/scala/weco/sierra/typesafe/SierraOauthHttpClientBuilder.scala
@@ -1,10 +1,10 @@
 package weco.sierra.typesafe
 
-import akka.actor.ActorSystem
-import akka.http.scaladsl.model.Uri
-import akka.http.scaladsl.model.headers.BasicHttpCredentials
+import org.apache.pekko.actor.ActorSystem
+import org.apache.pekko.http.scaladsl.model.Uri
+import org.apache.pekko.http.scaladsl.model.headers.BasicHttpCredentials
 import com.typesafe.config.Config
-import weco.http.client.{AkkaHttpClient, HttpGet, HttpPost}
+import weco.http.client.{HttpGet, HttpPost, PekkoHttpClient}
 import weco.sierra.http.SierraOauthHttpClient
 import weco.typesafe.config.builders.EnrichConfig._
 
@@ -19,7 +19,7 @@ object SierraOauthHttpClientBuilder {
     val username = config.requireString("sierra.api.key")
     val password = config.requireString("sierra.api.secret")
 
-    val client = new AkkaHttpClient() with HttpGet with HttpPost {
+    val client = new PekkoHttpClient() with HttpGet with HttpPost {
       override val baseUri: Uri = Uri(
         config.requireString("sierra.api.baseUrl")
       )