+++ typesafe_app/src/test/scala/weco/pekko/fixtures/Pekko.scala
@@ -1,13 +1,13 @@
-package weco.akka.fixtures
+package weco.pekko.fixtures
 
-import akka.actor.ActorSystem
-import akka.stream.Materializer
+import org.apache.pekko.actor.ActorSystem
+import org.apache.pekko.stream.Materializer
 import weco.fixtures._
 
 import scala.concurrent.Await
 import scala.concurrent.duration._
 
-trait Akka {
+trait Pekko {
   def withActorSystem[R]: Fixture[ActorSystem, R] = fixture[ActorSystem, R](
     create = ActorSystem(),
     destroy = system => Await.ready(system.terminate(), 10 seconds)